fix(dns): Enable LWIP DNS RFC6724 destination address selection to fix IPv6 TLS (#166)
TLS / HTTPS uses the hostname (for certificate validation) to directly look up
DNS. This change enables the RFC6724 destination address selection in LWIP DNS,
to return IPv6 addresses from DNS when a public IPv6 source is available, or
otherwise return IPv4.

i.e. Destination address selection is dynamci depending on what addresses are
available to send from.

Without this, LWIP is hard coded to prefer IPv4, which means when an IPv6-only
device tries to contact a dual-stack (or DNS64) server, it will get the IPv4
address (that can't be used). The config change fixes that.